Changed folder name. Fixed compile errors
