--- src/Runtime/Runtime/Empty Namespaces/System.Windows.Controls/WORKINPROGRESS/TabControl.cs
@@ -1,18 +1,0 @@
-﻿#if WORKINPROGRESS
-
-using System;
-using System.Windows.Controls.Primitives;
-using System.Windows.Documents;
-
-#if MIGRATION
-namespace System.Windows.Controls
-#else
-namespace Windows.UI.Xaml.Controls
-#endif
-{
-    public partial class TabControl : System.Windows.Controls.ItemsControl
-    {
-        public System.Windows.Controls.Dock TabStripPlacement { get; set; }
-    }
-}
-#endif
