+++ src/Runtime/Runtime/Empty Namespaces/Windows.UI.Xaml.Controls/WORKINPROGRESS/DataGridBoundColumn.cs
@@ -1,0 +1,16 @@
+﻿#if WORKINPROGRESS
+
+using System;
+
+#if MIGRATION
+namespace System.Windows.Controls
+#else
+namespace Windows.UI.Xaml.Controls
+#endif
+{
+    public abstract partial class DataGridBoundColumn : DataGridColumn
+    {
+        public Style ElementStyle { get; set; }
+    }
+}
+#endif
