+++ src/Runtime/Runtime/Empty Namespaces/Windows.UI.Xaml.Controls/WORKINPROGRESS/DataPicker.cs
@@ -1,7 +1,6 @@
 ﻿#if WORKINPROGRESS
 
 using System;
-using System.Windows.Documents;
 
 #if MIGRATION
 namespace System.Windows.Controls
@@ -9,11 +8,9 @@ namespace System.Windows.Controls
 namespace Windows.UI.Xaml.Controls
 #endif
 {
-    public partial class TextBlock : Control
+    public partial class DatePicker
     {
-        public FontSource FontSource { get; set; }
-
-        public static DependencyProperty LineHeightProperty
+        public static DependencyProperty CalendarStyleProperty
         {
             get; set;
         }
