+++ src/Runtime/Runtime/Empty Namespaces/Windows.UI.Xaml.Controls/WORKINPROGRESS/KeyEventArgs.cs
@@ -1,0 +1,6 @@
+﻿namespace Windows.UI.Xaml.Controls
+{
+    public class KeyEventArgs
+    {
+    }
+}
