+++ src/Runtime/Runtime/Empty Namespaces/Windows.UI.Xaml.Controls/WORKINPROGRESS/TabControl.cs
@@ -1,7 +1,6 @@
 ﻿#if WORKINPROGRESS
 
 using System;
-using System.Windows.Documents;
 
 #if MIGRATION
 namespace System.Windows.Controls
@@ -9,9 +8,9 @@ namespace System.Windows.Controls
 namespace Windows.UI.Xaml.Controls
 #endif
 {
-    public partial class HeaderedItemsControl : ItemsControl
+    public partial class TabControl : ItemsControl
     {
-        public object Header { get; set; }
+        public Dock TabStripPlacement { get; set; }
     }
 }
 #endif
