+++ src/Runtime/Runtime/Empty Namespaces/Windows.UI.Xaml.Input/WORKINPROGRESS/Cursors.cs
@@ -1,12 +1,9 @@
 ﻿#if WORKINPROGRESS
 
 using System;
+using System.Windows.Input;
 
-#if MIGRATION
 namespace System.Windows.Input
-#else
-namespace Windows.UI.Xaml.Input
-#endif
 {
     public static partial class Cursors
     {
