+++ src/Runtime/Runtime/Empty Namespaces/Windows.UI.Xaml/WORKINPROGRESS/Application.cs
