+++ src/Runtime/Runtime/Runtime.OpenSilver.csproj
@@ -49,33 +49,27 @@
 		<Compile Include="Core\Main\INTERNAL_NamedObject.cs" />
 		<Compile Include="Core\Main\WeakReferenceList.cs" />
 		<Compile Include="Core\Main\WeakReferenceListEnumerator.cs" />
-		<Compile Include="Empty Namespaces\System.Windows.Controls\WORKINPROGRESS\AutoCompleteBox.cs" />
-		<Compile Include="Empty Namespaces\System.Windows.Controls\WORKINPROGRESS\Calendar.cs" />
-		<Compile Include="Empty Namespaces\System.Windows.Controls\WORKINPROGRESS\Control.cs" />
-		<Compile Include="Empty Namespaces\System.Windows.Controls\WORKINPROGRESS\DataGridBoundColumn.cs" />
-		<Compile Include="Empty Namespaces\System.Windows.Controls\WORKINPROGRESS\DataPicker.cs" />
-		<Compile Include="Empty Namespaces\System.Windows.Controls\WORKINPROGRESS\HeaderedItemsControl.cs" />
-		<Compile Include="Empty Namespaces\System.Windows.Controls\WORKINPROGRESS\ISelectionAdapter.cs" />
-		<Compile Include="Empty Namespaces\System.Windows.Controls\WORKINPROGRESS\PopulatedEventArgs.cs" />
-		<Compile Include="Empty Namespaces\System.Windows.Controls\WORKINPROGRESS\SelectorSelectionAdapter.cs" />
-		<Compile Include="Empty Namespaces\System.Windows.Controls\WORKINPROGRESS\TabControl.cs" />
-		<Compile Include="Empty Namespaces\System.Windows.Controls\WORKINPROGRESS\TextBlock.cs" />
-		<Compile Include="Empty Namespaces\System.Windows.Controls\WORKINPROGRESS\WebBrowser.cs" />
-		<Compile Include="Empty Namespaces\System.Windows.Data\WORKINPROGRESS\Binding.cs" />
-		<Compile Include="Empty Namespaces\System.Windows.Documents\WORKINPROGRESS\FontSource.cs" />
-		<Compile Include="Empty Namespaces\System.Windows.Input\WORKINPROGRESS\Cursors.cs" />
-		<Compile Include="Empty Namespaces\System.Windows.Interactivity\WORKINPROGRESS\EventTrigger.cs" />
-		<Compile Include="Empty Namespaces\System.Windows.Printing\WORKINPROGRESS\PrintDocument.cs" />
-		<Compile Include="Empty Namespaces\System.Windows\WORKINPROGRESS\Application.cs" />
-		<Compile Include="Empty Namespaces\System.Windows\WORKINPROGRESS\VisualStateGroup.cs" />
-		<Compile Include="Empty Namespaces\System.Windows\WORKINPROGRESS\Window.cs" />
-		<Compile Include="Empty Namespaces\WORKINPROGRESS\System.Windows.Controls.cs" />
-		<Compile Include="Empty Namespaces\WORKINPROGRESS\System.Windows.cs" />
-		<Compile Include="Empty Namespaces\WORKINPROGRESS\System.Windows.Data.cs" />
-		<Compile Include="Empty Namespaces\WORKINPROGRESS\System.Windows.Documents.cs" />
-		<Compile Include="Empty Namespaces\WORKINPROGRESS\System.Windows.Input.cs" />
-		<Compile Include="Empty Namespaces\WORKINPROGRESS\System.Windows.Interactivity.cs" />
-		<Compile Include="Empty Namespaces\WORKINPROGRESS\System.Windows.Printing.cs" />
+		<Compile Include="Empty Namespaces\Windows.UI.Xaml.Controls\WORKINPROGRESS\AutoCompleteBox.cs" />
+		<Compile Include="Empty Namespaces\Windows.UI.Xaml.Controls\WORKINPROGRESS\Calendar.cs" />
+		<Compile Include="Empty Namespaces\Windows.UI.Xaml.Controls\WORKINPROGRESS\Control.cs" />
+		<Compile Include="Empty Namespaces\Windows.UI.Xaml.Controls\WORKINPROGRESS\DataGridBoundColumn.cs" />
+		<Compile Include="Empty Namespaces\Windows.UI.Xaml.Controls\WORKINPROGRESS\DataPicker.cs" />
+		<Compile Include="Empty Namespaces\Windows.UI.Xaml.Controls\WORKINPROGRESS\HeaderedItemsControl.cs" />
+		<Compile Include="Empty Namespaces\Windows.UI.Xaml.Controls\WORKINPROGRESS\ISelectionAdapter.cs" />
+		<Compile Include="Empty Namespaces\Windows.UI.Xaml.Controls\WORKINPROGRESS\KeyEventArgs.cs" />
+		<Compile Include="Empty Namespaces\Windows.UI.Xaml.Controls\WORKINPROGRESS\PopulatedEventArgs.cs" />
+		<Compile Include="Empty Namespaces\Windows.UI.Xaml.Controls\WORKINPROGRESS\SelectorSelectionAdapter.cs" />
+		<Compile Include="Empty Namespaces\Windows.UI.Xaml.Controls\WORKINPROGRESS\TabControl.cs" />
+		<Compile Include="Empty Namespaces\Windows.UI.Xaml.Controls\WORKINPROGRESS\TextBlock.cs" />
+		<Compile Include="Empty Namespaces\Windows.UI.Xaml.Controls\WORKINPROGRESS\WebBrowser.cs" />
+		<Compile Include="Empty Namespaces\Windows.UI.Xaml.Data\WORKINPROGRESS\Binding.cs" />
+		<Compile Include="Empty Namespaces\Windows.UI.Xaml.Documents\WORKINPROGRESS\FontSource.cs" />
+		<Compile Include="Empty Namespaces\Windows.UI.Xaml.Input\WORKINPROGRESS\Cursors.cs" />
+		<Compile Include="Empty Namespaces\Windows.UI.Xaml.Interactivity\WORKINPROGRESS\EventTrigger.cs" />
+		<Compile Include="Empty Namespaces\Windows.UI.Xaml.Printing\WORKINPROGRESS\PrintDocument.cs" />
+		<Compile Include="Empty Namespaces\Windows.UI.Xaml\WORKINPROGRESS\Application.cs" />
+		<Compile Include="Empty Namespaces\Windows.UI.Xaml\WORKINPROGRESS\VisualStateGroup.cs" />
+		<Compile Include="Empty Namespaces\Windows.UI.Xaml\WORKINPROGRESS\Window.cs" />
 		<Compile Include="Microsoft.Expression.Interactivity.Core\WORKINPROGRESS\ChangePropertyAction.cs" />
 		<Compile Include="Microsoft.Expression.Interactivity.Core\WORKINPROGRESS\ComparisonConditionType.cs" />
 		<Compile Include="Microsoft.Expression.Interactivity.Core\WORKINPROGRESS\ControlStoryboardAction.cs" />
